Add line ending insensitive comparisons
Replace the \r\n and \r line endings with \n to get insensitive comparisons.

Add LineEndingInsensitiveString class and associated DifferenceHighlighter
to get the difference highlighting output. The ShouldlyAssertionContext
now wraps the expected and actual in a LineEndingInsensitiveString if the
IgnoreLineEndings option is set.

Add tests for LineEndingInsensitiveStringDifferenceHighlighter,
NormalizeLineEndings extension method, and the new behavior for
ShouldBe.